% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connect.R
\name{get_plattdeutsch}
\alias{get_plattdeutsch}
\title{get_plattdeutsch}
\usage{
get_plattdeutsch(txt = "", random = F)
}
\arguments{
\item{txt}{provide a text you want to translate from German into Plattdeutsch (a maximum of 219 characters per call)}

\item{random}{provides a random text in Plattdeutsch}
}
\description{
Translates German to Plattdeuts
}
